requirements.txt
django==5.1
djangorestframework==3.15.2
pytest-django==4.9.0
Faker>=19.3.0
requests>=2.31.0
Pillow>=10.0.0